backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
python-docx==1.2.0
reportlab==4.4.3
httpx==0.28.1
youtube-transcript-api==1.2.3
yt-dlp
pytesseract
openai
